Fix broken merge
Whoops. Bad merge with the csprojs leading to including both old and new in one file.
Convert tabs to spaces and remove and sort usings

What it says in the title. We started it with the .Net core changes so let us keep going.
Readd source link

That merge was bad